+++ src/folder/mod.rs
@@ -2,9 +2,12 @@ pub mod arg;
 pub mod command;
 pub mod config;
 
-use comfy_table::{presets, Attribute, Cell, ContentArrangement, Row, Table};
+use comfy_table::{Cell, ContentArrangement, Row, Table};
+use crossterm::style::Color;
 use serde::{Serialize, Serializer};
 use std::{fmt, ops::Deref};
+
+use self::config::ListFoldersTableConfig;
 
 #[derive(Clone, Debug, Default, Serialize)]
 pub struct Folder {
@@ -13,11 +16,12 @@ pub struct Folder {
 }
 
 impl Folder {
-    pub fn to_row(&self) -> Row {
+    pub fn to_row(&self, config: &ListFoldersTableConfig) -> Row {
         let mut row = Row::new();
+        row.max_height(1);
 
-        row.add_cell(Cell::new(&self.name).fg(comfy_table::Color::Blue));
-        row.add_cell(Cell::new(&self.desc).fg(comfy_table::Color::Green));
+        row.add_cell(Cell::new(&self.name).fg(config.name_color()));
+        row.add_cell(Cell::new(&self.desc).fg(config.desc_color()));
 
         row
     }
@@ -34,23 +38,6 @@ impl From<email::folder::Folder> for Folder {
 
 #[derive(Clone, Debug, Default, Serialize)]
 pub struct Folders(Vec<Folder>);
-
-impl Folders {
-    pub fn to_table(&self) -> Table {
-        let mut table = Table::new();
-
-        table
-            .load_preset(presets::NOTHING)
-            .set_content_arrangement(ContentArrangement::DynamicFullWidth)
-            .set_header(Row::from([
-                Cell::new("NAME").add_attribute(Attribute::Reverse),
-                Cell::new("DESC").add_attribute(Attribute::Reverse),
-            ]))
-            .add_rows(self.iter().map(Folder::to_row));
-
-        table
-    }
-}
 
 impl Deref for Folders {
     type Target = Vec<Folder>;
@@ -69,11 +56,27 @@ impl From<email::folder::Folders> for Folders {
 pub struct FoldersTable {
     folders: Folders,
     width: Option<u16>,
+    config: ListFoldersTableConfig,
 }
 
 impl FoldersTable {
     pub fn with_some_width(mut self, width: Option<u16>) -> Self {
         self.width = width;
+        self
+    }
+
+    pub fn with_some_preset(mut self, preset: Option<String>) -> Self {
+        self.config.preset = preset;
+        self
+    }
+
+    pub fn with_some_name_color(mut self, color: Option<Color>) -> Self {
+        self.config.name_color = color;
+        self
+    }
+
+    pub fn with_some_desc_color(mut self, color: Option<Color>) -> Self {
+        self.config.desc_color = color;
         self
     }
 }
@@ -83,13 +86,24 @@ impl From<Folders> for FoldersTable {
         Self {
             folders,
             width: None,
+            config: Default::default(),
         }
     }
 }
 
 impl fmt::Display for FoldersTable {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let mut table = self.folders.to_table();
+        let mut table = Table::new();
+
+        table
+            .load_preset(self.config.preset())
+            .set_content_arrangement(ContentArrangement::DynamicFullWidth)
+            .set_header(Row::from([Cell::new("NAME"), Cell::new("DESC")]))
+            .add_rows(
+                self.folders
+                    .iter()
+                    .map(|folder| folder.to_row(&self.config)),
+            );
 
         if let Some(width) = self.width {
             table.set_width(width);