+++ src/ui/mod.rs
@@ -1,3 +1,29 @@
+use crossterm::style::Color;
+
 pub mod choice;
 pub mod editor;
 pub(crate) mod prompt;
+
+pub(crate) fn map_color(color: Color) -> comfy_table::Color {
+    match color {
+        Color::Reset => comfy_table::Color::Reset,
+        Color::Black => comfy_table::Color::Black,
+        Color::DarkGrey => comfy_table::Color::DarkGrey,
+        Color::Red => comfy_table::Color::Red,
+        Color::DarkRed => comfy_table::Color::DarkRed,
+        Color::Green => comfy_table::Color::Green,
+        Color::DarkGreen => comfy_table::Color::DarkGreen,
+        Color::Yellow => comfy_table::Color::Yellow,
+        Color::DarkYellow => comfy_table::Color::DarkYellow,
+        Color::Blue => comfy_table::Color::Blue,
+        Color::DarkBlue => comfy_table::Color::DarkBlue,
+        Color::Magenta => comfy_table::Color::Magenta,
+        Color::DarkMagenta => comfy_table::Color::DarkMagenta,
+        Color::Cyan => comfy_table::Color::Cyan,
+        Color::DarkCyan => comfy_table::Color::DarkCyan,
+        Color::White => comfy_table::Color::White,
+        Color::Grey => comfy_table::Color::Grey,
+        Color::Rgb { r, g, b } => comfy_table::Color::Rgb { r, g, b },
+        Color::AnsiValue(n) => comfy_table::Color::AnsiValue(n),
+    }
+}